feat: dashboard redesign, subscription overhaul, payment improvements, and UI cleanup
Admin frontend:
- New dashboard components: activity stream, bento tools, HUD metrics, pulse header
- Remove deprecated pages: affiliates, subscriptions, debug/auth, manual, profile, docs/api
- Redesign access-control plans editor, wallet management, notifications, payments
- Overhaul auth page, settings dashboard, error boundaries
- New sidebar component with updated navigation

Frontend:
- Payment flow: enhanced confirmation step, plan comparison card, upgrade banner
- Chat: improved inbox, panel, topic selector with better UX
- Portfolio: progressive auth integration, watchlist provider improvements
- Home: updated hero section and dynamic pricing layout

Backend:
- Subscription handlers: improved upgrade/downgrade flows
- Audit log repository refactoring
- Payment credit handlers and submit tx improvements

Shared:
- API client updates across admin, analytics, auth, credits modules
- Auth client: improved token handling and cookie management
- Notification bell and SSE notifications refactoring
- Shared components: pricing card, chat markdown, stock data card redesigns
- Logger and tier-helpers cleanup
- Synpress E2E test setup and contracts documentation

+++ apps/backend/src/web/payments/credit_handlers.rs
@@ -211,7 +211,7 @@ pub async fn admin_grant_credits(
     info!("Admin {} granting {} credits to {}", admin_wallet, request.amount, request.wallet_address);
 
     // Validate amount is positive
-    if request.amount <= BigDecimal::from(0) {
+    if request.amount <= 0 {
         return Err(Json(UnifiedErrorResponse::new(400, "Invalid amount", "Amount must be positive")));
     }
 
@@ -301,7 +301,7 @@ pub async fn admin_revoke_credits(
     info!("Admin {} revoking {} credits from {}", admin_wallet, request.amount, request.wallet_address);
 
     // Validate amount is positive
-    if request.amount <= BigDecimal::from(0) {
+    if request.amount <= 0 {
         return Err(Json(UnifiedErrorResponse::new(400, "Invalid amount", "Amount must be positive")));
     }
